Refactoring. Move new entities to domain package
